Project2/requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
blind-watermark==0.4.0 